commit storybook init and early progress
This commit only sets up storybook and attempts to add the combobox.jsx component.  But runs into an issue with the combobox that I can't quite figure out.

Leaving this branch at this point to work on utilizing Radix UI, and will hope to replace components with Radix components and hope they work easier with storybook.