Place IFrame and HeaderNav side by side
Moves the iframe and the headerNav into a 'row' div which is styled with flex, so the elements are side by side.  This is a choice to have the header navigation not overlap the brew when it's open, but instead squish that iframe area but keep it fully visible (the zoom to fit button can be used to resize the page if needed).