set styling for the `.row` div
This sets the HeaderNav and Iframe row to have flex styling (side by side).  Likely more could be done here, such as flex-wrap if on a very small screen so that the toc is on top or something, but not worrying about that now.

I could likely get away with targeting just `.content` and skip `.pane` but this felt more explicit (/share/ doesn't have `.pane`).