Put HeaderNav button in Toolbar
HeaderNav toggle goes into Toolbar, and small changes to state variables which keep track of whether it is open or not in order to change classNames.  Toolbar is reconfigured as well to have "start" and "end" groups which push their children to the start or end of the toolbar rather than centering.

add state variable