Add input validation
Takes validator functions as props.   Validation is done on the current input, for each tag + the "new tag input", rather than on the TagInput component as a whole.  Each validator function (there can be an array of them) is checked against the input value.  If it fails, it triggers setCustomValidity method on the HTMLElement with a custom message.

On the next keypress, the validity is refreshed until it is submitted again.