Add a check for uniqueness among values
In the validation step, check that the new value is unique among tags/values.  If not, add an error to the list of validation errors, which is then added to the error message with the rest of the errors.

Checks for uniqueness by default, but requireUnique prop can be set to false to skip in case it's needed.